# postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Intern Kominfo API
  version: 1.1.0
  description: |
    API for the Intern Kominfo application — a government intern management
    platform with Work-From-Home (WFH) reporting, change management workflows,
    user administration, and QR-based document verification.

    Authentication uses Laravel Sanctum bearer tokens.
    Authorization uses Spatie permission-based roles (admin, kepala_tim, staf).
servers:
  - url: http://localhost
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Email atau password salah.
...
    StoreImplementationRequest:
      type: object
      properties:
        priority:
          type: string
          enum: [low, medium, high, critical]
          nullable: true
          example: high
        impact:
          type: string
          enum: [low, medium, high]
          nullable: true
          example: high
        production_impact:
          type: string
          nullable: true
          example: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort:
          type: string
          nullable: true
          example: "3 orang x 5 hari kerja"
        cost_needed:
          type: boolean
          nullable: true
          example: true
        cost_amount:
          type: number
          format: double
          nullable: true
          example: 15000000
        resources:
          type: string
          nullable: true
          example: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan:
          type: string
          nullable: true
          example: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        change_type_ids:
          type: array
          nullable: true
          items:
            type: integer
          example: [1, 3]
        implementation_result:
          type: string
          nullable: true
          example: null
        testing_result:
          type: string
          nullable: true
          example: null
      example:
        priority: high
        impact: high
        production_impact: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort: "3 orang x 5 hari kerja"
        cost_needed: true
        cost_amount: 15000000
        resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        change_type_ids: [1, 3]
    ImplementationResource:
      type: object
      required: [id, change_initiation_id, status]
      properties:
        id:
          type: integer
          example: 4
        change_initiation_id:
          type: integer
          example: 8
        priority:
          type: string
          nullable: true
          example: high
        impact:
          type: string
          nullable: true
          example: high
        production_impact:
          type: string
          nullable: true
          example: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort:
          type: string
          nullable: true
          example: "3 orang x 5 hari kerja"
        cost_needed:
          type: boolean
          nullable: true
          example: true
        cost_amount:
          type: number
          format: double
          nullable: true
          example: 15000000
        resources:
          type: string
          nullable: true
          example: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan:
          type: string
          nullable: true
          example: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        evaluator_id:
          type: integer
          nullable: true
          example: 5
        evaluator_signed_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-07-14T11:00:00.000000Z"
        review_status:
          type: string
          nullable: true
          example: null
        review_response:
          type: string
          nullable: true
          example: null
        execution_date:
          type: string
          format: date
          nullable: true
          example: null
        reviewer_id:
          type: integer
          nullable: true
          example: null
        reviewer_signed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        implementation_result:
          type: string
          nullable: true
          example: null
        testing_result:
          type: string
          nullable: true
          example: null
        release_date:
          type: string
          format: date
          nullable: true
          example: null
        responsible_id:
          type: integer
          nullable: true
          example: null
        responsible_signed_at:
          type: string
          format: date-time
          nullable: true
          example: null
        status:
          type: string
          enum: [draft, submitted, completed, rejected, revision]
          example: submitted
        change_types:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 1
              name:
                type: string
                example: Perubahan Infrastruktur
        attachments:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 15
              path:
                type: string
                example: implementations/4/diagram-arsitektur.png
              url:
                type: string
                format: uri
                example: http://localhost/storage/implementations/4/diagram-arsitektur.png
              sort_order:
                type: integer
                example: 1
      example:
        id: 4
        change_initiation_id: 8
        priority: high
        impact: high
        production_impact: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
        required_effort: "3 orang x 5 hari kerja"
        cost_needed: true
        cost_amount: 15000000
        resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
        test_plan: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
        evaluator_id: 5
        evaluator_signed_at: "2025-07-14T11:00:00.000000Z"
        review_status: null
        review_response: null
        execution_date: null
        reviewer_id: null
        reviewer_signed_at: null
        implementation_result: null
        testing_result: null
        release_date: null
        responsible_id: null
        responsible_signed_at: null
        status: submitted
        change_types:
          - id: 1
            name: Perubahan Infrastruktur
          - id: 3
            name: Perubahan Data
        attachments:
          - id: 15
            path: implementations/4/diagram-arsitektur.png
            url: http://localhost/storage/implementations/4/diagram-arsitektur.png
            sort_order: 1
...
  parameters:
...
    InitiationIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Change Initiation ID
      example: 8
    ImplementationIdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Change Implementation ID
      example: 4
security:
  - bearerAuth: []
...
paths:
...
  /api/changes/initiations/{id}/implementations:
    post:
      tags: [Change Implementation]
      summary: Create an implementation under an initiation
      operationId: createImplementation
      parameters:
        - $ref: "#/components/parameters/InitiationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreImplementationRequest"
            example:
              priority: high
              impact: high
              production_impact: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
              required_effort: "3 orang x 5 hari kerja"
              cost_needed: true
              cost_amount: 15000000
              resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
              test_plan: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
              change_type_ids: [1, 3]
      responses:
        "201":
          description: Implementation created
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Implementasi perubahan berhasil dibuat.
                  data:
                    $ref: "#/components/schemas/ImplementationResource"
              example:
                success: true
                message: Implementasi perubahan berhasil dibuat.
                data:
                  id: 4
                  change_initiation_id: 8
                  priority: high
                  impact: high
                  production_impact: "Aplikasi akan tidak dapat diakses selama 2 jam pada saat migrasi data."
                  required_effort: "3 orang x 5 hari kerja"
                  cost_needed: true
                  cost_amount: 15000000
                  resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps"
                  test_plan: "Smoke test semua endpoint API, load test dengan 100 concurrent users, uji rollback ke server lama."
                  evaluator_id: null
                  evaluator_signed_at: null
                  review_status: null
                  review_response: null
                  execution_date: null
                  reviewer_id: null
                  reviewer_signed_at: null
                  implementation_result: null
                  testing_result: null
                  release_date: null
                  responsible_id: null
                  responsible_signed_at: null
                  status: draft
                  change_types:
                    - id: 1
                      name: Perubahan Infrastruktur
                    - id: 3
                      name: Perubahan Data
                  attachments: []
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Missing permission change.implementation.create
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
        "404":
          description: Initiation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Inisiasi tidak ditemukan.
        "422":
          description: Initiation not approved or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: "Inisiasi belum disetujui, tidak dapat membuat implementasi."
  /api/changes/implementations/{id}:
    get:
      tags: [Change Implementation]
      summary: Get implementation by ID
      operationId: getImplementationById
      parameters:
        - $ref: "#/components/parameters/ImplementationIdPath"
      responses:
        "200":
          description: Implementation data
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/ImplementationResource"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Not evaluator and missing permission change.implementation.review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
        "404":
          description: Implementation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Implementasi tidak ditemukan.
    put:
      tags: [Change Implementation]
      summary: Update implementation
      operationId: updateImplementation
      parameters:
        - $ref: "#/components/parameters/ImplementationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreImplementationRequest"
            example:
              priority: critical
              impact: high
              production_impact: "Aplikasi akan tidak dapat diakses selama 3 jam pada saat migrasi data."
              required_effort: "4 orang x 5 hari kerja"
              cost_needed: true
              cost_amount: 18000000
              resources: "Cloud server 8 vCPU 16GB RAM, storage 200GB SSD, bandwidth 1Gbps, backup server"
              test_plan: "Smoke test semua endpoint API, load test dengan 200 concurrent users, uji rollback ke server lama."
              change_type_ids: [1, 3]
      responses:
        "200":
          description: Implementation updated
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Implementasi perubahan berhasil diperbarui.
                  data:
                    $ref: "#/components/schemas/ImplementationResource"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Not evaluator or missing permission change.implementation.update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
        "404":
          description: Implementation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Implementasi tidak ditemukan.
        "422":
          description: Not in draft/revision status or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: "Implementasi dengan status completed tidak dapat diperbarui."
...
  /api/changes/implementations/{id}/submit:
    post:
      tags: [Change Implementation]
      summary: Submit implementation for review
      operationId: submitImplementation
      parameters:
        - $ref: "#/components/parameters/ImplementationIdPath"
      responses:
        "200":
          description: Implementation submitted
          content:
            application/json:
              schema:
                type: object
                required: [success, message, data]
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Implementasi berhasil disubmit untuk review.
                  data:
                    $ref: "#/components/schemas/ImplementationResource"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Unauthenticated.
        "403":
          description: Not evaluator or missing permission change.implementation.submit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: This action is unauthorized.
        "404":
          description: Implementation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: Implementasi tidak ditemukan.
        "422":
          description: Not in draft/revision status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                message: "Implementasi dengan status completed tidak dapat disubmit."
...